company table fix
